move(GetDrinkPreviewResponse) : move to drink directory
