+++ src/main/java/kuit/project/beering/dto/response/drink/GetDrinkPreviewResponseBuilder.java
@@ -1,7 +1,6 @@
-package kuit.project.beering.dto.response.favorite;
+package kuit.project.beering.dto.response.drink;
 
 import kuit.project.beering.domain.Drink;
-import kuit.project.beering.dto.response.ResponseBuilder;
 
 public class GetDrinkPreviewResponseBuilder{
