+++ src/main/java/kuit/project/beering/service/DrinkService.java
@@ -7,8 +7,8 @@ import kuit.project.beering.domain.image.Image;
 import kuit.project.beering.dto.request.drink.DrinkSearchCondition;
 import kuit.project.beering.dto.request.drink.SearchDrinkRequest;
 import kuit.project.beering.dto.request.drink.SortType;
-import kuit.project.beering.dto.response.favorite.GetDrinkPreviewResponse;
-import kuit.project.beering.dto.response.favorite.GetDrinkPreviewResponseBuilder;
+import kuit.project.beering.dto.response.drink.GetDrinkPreviewResponse;
+import kuit.project.beering.dto.response.drink.GetDrinkPreviewResponseBuilder;
 import kuit.project.beering.dto.response.drink.*;
 import kuit.project.beering.repository.FavoriteRepository;
 import kuit.project.beering.repository.MemberRepository;
